API вынесены в отдельный модуль + добавлены аннотации + генерация email через Faker
Для генерации email аннотацию не добавлял, тк это чисто вспомогательный метод, возвращающий одну строку
