+++ requirements.txt
@@ -1,3 +1,5 @@
 allure-pytest==2.14.1
 pytest==8.3.5
 requests==2.32.3
+allure-python-commons~=2.14.1
+Faker==37.3.0
